Add polling to dev server warm-up.
